+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `title` parameter to `/api/v1/statuses` form data.
+- Add `title` field to `Status` API entity if ActivityPub object has a `name`.
+
 ### Changed
 
 - Changed MSRV to 1.82.0.